+++ src/types/types.rs
@@ -1,0 +1,4 @@
+// TODO: the IDs should be u128
+pub type StreamId = u64;
+pub type CheckBlockId = u64;
+pub type BlockIndex = usize;
